Add support for multiple artists in clicking
Basically, it splits the artist field by commas and ampersands
and makes each individual one clickable. Also, it does the same
thing with songs that have "feat." in the title.

Not bad, eh? It is definitely a hack, but it works well.
Effectively closes #7.

Also colours scrollbars, so closes #29.
Also fixes flatly theme, so closes #28.