Add "add to new playlist" feature and fix updates for play/pause
Closes #30:

Simple feature, really: have a "add to new playlist" to add the
selected songs to a new playlist. A dialog box pops up first to
ask the user what the playlist name should be, then makes the
playlist and adds the songs to it.

Future TODO's: eliminate code reuse between the "Add playlist"
button and the "new playlist" button.

Closes #33:

The UI should now synchonize with the background audio element.
Tested with Safari 13 on macOS Mojave, ymmv with other setups
(then again, not a lot of browers let you control the background
audio element, so this is probably a non-issue for most people)

(not an issue, but still a nice new feature):

Pause the 50ms interval when the player is not running. Not sure
why it kept running in the first place but now it's paused.
Hopefully this reduces the battery usage or something like that...