Update heroku build (Mar 2020) (#34)
* use https lastfm api

* Update dependencies (#20)

* fail at trying to update deps

* get rid of bower, but keep grunt

* move themes into gruntfile, clena up gruntfile

* add default config

* Overhaul theme system

Bootswatch themes are now properly included, with the sliders
looking nice, too :)

Closes #15

* Add old themes that I made like 4 years ago

Closes #22

* Don't set duration to 0 when single-repeating and chromecasting

The media reference has been invalidated at this point so the Seek
Request fails. Instead, load up the same song again and play it.

Closes #24.

* Add check for undefined room

Looks like NAMENOTFOUND was already coded in, but this error
was preventing it from being returned.

Closes #25.

* Add support for multiple artists in clicking

Basically, it splits the artist field by commas and ampersands
and makes each individual one clickable. Also, it does the same
thing with songs that have "feat." in the title.

Not bad, eh? It is definitely a hack, but it works well.
Effectively closes #7.

Also colours scrollbars, so closes #29.
Also fixes flatly theme, so closes #28.

* Try electron building, but it doesn't work

* Raise volume bar

Closes #27

Honestly the bar is a bit high now, but that's ok i guess...

* Fix issue with some formats of featuring artists

* Fix playlist downloading

* Update package dependencies

* update library_functions.js for new version of async.waterfall

Fixes #32.

* fix youtube downloading with newest versions of deps

* Fix code issues caused by newest dependency update

mkdirp was updated with breaking changes and ytdl-core seemed to
lose a few API options.

* continue fixes from updated mkdirp version

* Add "add to new playlist" feature and fix updates for play/pause

Closes #30:

Simple feature, really: have a "add to new playlist" to add the
selected songs to a new playlist. A dialog box pops up first to
ask the user what the playlist name should be, then makes the
playlist and adds the songs to it.

Future TODO's: eliminate code reuse between the "Add playlist"
button and the "new playlist" button.

Closes #33:

The UI should now synchonize with the background audio element.
Tested with Safari 13 on macOS Mojave, ymmv with other setups
(then again, not a lot of browers let you control the background
audio element, so this is probably a non-issue for most people)

(not an issue, but still a nice new feature):

Pause the 50ms interval when the player is not running. Not sure
why it kept running in the first place but now it's paused.
Hopefully this reduces the battery usage or something like that...